feat: authorize the update and delete Project routes

+++ crm-api/app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\CreateProjectRequest;
 use App\Http\Requests\UpdateProjectRequest;
 use App\Http\Resources\ProjectResource;
 use App\Models\Project;
+use App\Models\Team;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -60,8 +61,16 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project)
     {
-        $project->update($request->validated());
-        return ProjectResource::make($project);
+        $team = Team::firstWhere('project_id', $project->id);
+        $user = $request->user();
+        if ($user->owns($project) || $user->isAbleTo('project-update', $team)) {
+            $project->update($request->validated());
+            return ProjectResource::make($project);
+        }
+        abort(
+            Response::HTTP_FORBIDDEN,
+            "You don't have the correct permissions for this action"
+        );
     }
 
     /**
@@ -70,12 +79,20 @@ class ProjectController extends Controller
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Project $project)
+    public function destroy(Request $request, Project $project)
     {
-        $project->delete();
-        return response()->json([
-            'message' => 'Project Deleted',
-            Response::HTTP_NO_CONTENT
-        ]);
+        $user = $request->user();
+        $team = Team::firstWhere('project_id', $project->id);
+        if ($user->owns($project) || $user->hasRole('moderator', $team) || $user->isAbleTo('project-delete', $team)) {
+            $project->delete();
+            return response()->json([
+                'message' => 'Project Deleted',
+                Response::HTTP_NO_CONTENT
+            ]);
+        }
+        abort(
+            Response::HTTP_FORBIDDEN,
+            "You don't have the correct permissions for this action"
+        );
     }
 }
